+++ packages/web3-core/test/unit/web3_subscription_old_providers.test.ts
@@ -17,10 +17,13 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { ExampleSubscription } from './fixtures/example_subscription';
 import { Web3EventEmitter } from '../../src/web3_event_emitter';
+import { Web3SubscriptionManager } from '../../src';
 
 describe('Web3Subscription', () => {
 	let requestManager: any;
+	let subscriptionManager: Web3SubscriptionManager<any, any>;
 	let eipRequestManager: any;
+	let subscriptionManagerWithEipReqMan: Web3SubscriptionManager<any, any>;
 	let provider: Web3EventEmitter<any>;
 	let eipProvider: Web3EventEmitter<any>;
 
@@ -29,21 +32,37 @@ describe('Web3Subscription', () => {
 		eipProvider = new Web3EventEmitter();
 		// @ts-expect-error add to test eip providers
 		eipProvider.request = jest.fn();
-		requestManager = { send: jest.fn(), on: jest.fn(), provider };
-		eipRequestManager = { send: jest.fn(), on: jest.fn(), provider: eipProvider };
+		requestManager = {
+			send: jest.fn().mockImplementation(async () => {
+				return 'sub-id';
+			}),
+			on: jest.fn(),
+			provider,
+		};
+		subscriptionManager = new Web3SubscriptionManager(requestManager, {});
+
+		eipRequestManager = {
+			send: jest.fn().mockImplementation(async () => {
+				return 'sub-id';
+			}),
+			on: jest.fn(),
+			provider: eipProvider,
+		};
+		subscriptionManagerWithEipReqMan = new Web3SubscriptionManager(eipRequestManager, {});
 	});
 
 	describe('providers response for old provider', () => {
 		it('data with result', async () => {
+			const sub = new ExampleSubscription({ param1: 'param1' }, { subscriptionManager });
+			await sub.subscribe();
 			const testData = {
 				data: {
+					subscription: sub.id,
 					result: {
 						some: 1,
 					},
 				},
 			};
-			const sub = new ExampleSubscription({ param1: 'param1' }, { requestManager });
-			await sub.subscribe();
 			// @ts-expect-error spy on protected method
 			const processResult = jest.spyOn(sub, '_processSubscriptionResult');
 			provider.emit('data', testData);
@@ -51,33 +70,35 @@ describe('Web3Subscription', () => {
 		});
 
 		it('data without result for old provider', async () => {
+			const sub = new ExampleSubscription({ param1: 'param1' }, { subscriptionManager });
+			await sub.subscribe();
 			const testData = {
 				data: {
+					subscription: sub.id,
 					other: {
 						some: 1,
 					},
 				},
 			};
-			const sub = new ExampleSubscription({ param1: 'param1' }, { requestManager });
-			await sub.subscribe();
 			// @ts-expect-error spy on protected method
 			const processResult = jest.spyOn(sub, '_processSubscriptionResult');
 			provider.emit('data', testData);
 			expect(processResult).toHaveBeenCalledWith(testData.data);
 		});
 		it('data with result for eipProvider', async () => {
+			const sub = new ExampleSubscription(
+				{ param1: 'param1' },
+				{ subscriptionManager: subscriptionManagerWithEipReqMan },
+			);
+			await sub.subscribe();
 			const testData = {
 				data: {
+					subscription: sub.id,
 					result: {
 						some: 1,
 					},
 				},
 			};
-			const sub = new ExampleSubscription(
-				{ param1: 'param1' },
-				{ requestManager: eipRequestManager },
-			);
-			await sub.subscribe();
 			// @ts-expect-error spy on protected method
 			const processResult = jest.spyOn(sub, '_processSubscriptionResult');
 			eipProvider.emit('message', testData);
@@ -85,18 +106,19 @@ describe('Web3Subscription', () => {
 		});
 
 		it('data without result for eipProvider', async () => {
+			const sub = new ExampleSubscription(
+				{ param1: 'param1' },
+				{ subscriptionManager: subscriptionManagerWithEipReqMan },
+			);
+			await sub.subscribe();
 			const testData = {
 				data: {
+					subscription: sub.id,
 					other: {
 						some: 1,
 					},
 				},
 			};
-			const sub = new ExampleSubscription(
-				{ param1: 'param1' },
-				{ requestManager: eipRequestManager },
-			);
-			await sub.subscribe();
 			// @ts-expect-error spy on protected method
 			const processResult = jest.spyOn(sub, '_processSubscriptionResult');
 			eipProvider.emit('message', testData);