+++ packages/web3-eth-contract/CHANGELOG.md
@@ -272,8 +272,10 @@ Documentation:
 [Web3 API](https://docs.web3js.org/api)
 [Migration Guide from 1.x](https://docs.web3js.org/guides/web3_upgrade_guide/x/)
 
-## [Unreleased]
+## [4.0.2]
 
 ### Fixed
 
 -   Event filtering using non-indexed and indexed string event arguments (#6167)
+
+## [Unreleased]